file re-organization, split the huge main file into multiple smaller files
