+++ test/runtests.jl
@@ -1,10 +1,12 @@
-tests = ["viewtypes",  
-         "contrank", 
-         "subviews"]
+tests = ["contiguousviews",
+         "stridedviews",
+         "contrank",
+         "subviews",
+         "linalg",
+         "convenience"]
 
 for t in tests
     fp = string(t, ".jl")
     println("* running $fp ...")
     include(fp)
 end
-
